+++ include/rule.h
@@ -21,7 +21,7 @@ struct rule {
 	Pvoid_t		stores_JQ; /* (uint64_t seq) -> (struct fref *store) */
 	Pvoid_t		copies_JQ; /* (uint64_t seq) -> (struct fref *copy) */
 	Pvoid_t		writes_JQ; /* (uint64_t seq) -> (struct fval *wrt) */
-	size_t		refcnt;
+	uint32_t	refcnt;
 };
